+++ app/assets/javascripts/app/helpers/text_formatter.js
@@ -43,7 +43,7 @@
 
     var hashtagPlugin = window.markdownitHashtag;
     md.use(hashtagPlugin, {
-      // compare tag_text_regexp in app/models/acts_as_taggable_on-tag.rb
+      // compare tag_text_regexp in config/initializers/acts_as_taggable_on.rb
       hashtagRegExp: "[" + PosixBracketExpressions.word +
                            "\\u055b" + // Armenian emphasis mark
                            "\\u055c" + // Armenian exclamation mark
